+++ src/grafanarmadillo/templator.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 
 import logging
 from pathlib import Path
-from typing import Callable, Dict, NewType, Optional
+from typing import Any, Callable, Dict, NewType, Optional
 
 from grafana_client.client import GrafanaClientError
 
@@ -103,7 +103,7 @@ class Templator:
 		"""Convert a dashboard into a one ready for templating."""
 		new = dashboard.copy()
 		new = project_dict(
-			new, {"id", "uid"}, inverse=True
+			new, {"id", "uid", "folderUID"}, inverse=True
 		)  # we don't erase the title so that we can template it later
 
 		return self.make_template(DashboardContent(new))
@@ -146,6 +146,18 @@ def make_mapping_templator(mapping: EnvMapping, env_grafana: str, env_template: 
 	template_to_grafana = {v: k for k, v in grafana_to_template.items()}
 
 	return Templator(make_template=findreplace(grafana_to_template), fill_template=findreplace(template_to_grafana))
+
+
+def fill_grafana_templating_options(options: dict[str, Any]) -> DashboardTransformer:
+	"""Fill the options on a Grafana template (the kind of dashboard available from grafana.com)."""
+	def _fill_remote_options(d: DashboardContent) -> DashboardContent:
+		for k, v in options.items():
+			for e in d["templating"]["list"]:
+				if e["name"] == k:
+					e["current"] = v
+		return d
+
+	return _fill_remote_options
 
 
 def alert_dashboarduid_templator(finder: Finder) -> Templator: