+++ docs/CHANGELOG.rst
@@ -2,6 +2,24 @@
 Changelog
 =========
 
+* task : upgrade pinned deps
+* task : (breaking) drop support for Python 3.8
+* fix : also remove folderUID when creating templates
+* feature : caching in finder
+* feature : `FlowResult.ensure_success` for inline raising flow exceptions
+* feature : `FileStore.resolve_object_to_filepath` has the type of object being referenced, allowing separate folders for alerts and dashboards
+* feature : `Finder.get_alerts_in_folders`
+* refactor : (breaking) set compatibility for API in Finder
+* refactor : use `folderUIDs` for newer search API
+* feature : FileStore accepts json encoders and decoders
+* feature : Flow.extend for adding multiple flows in 1 call
+* fix : searching for multiple folders doesn't fail and return all of them
+* feature : add DashboardTransformer to help fill Grafana templates
+* feature : flow: add URLStore to import dashboards and alerts from remote sources
+* feature : flow: allow all PathLike instead of just str
+* feature : flow: easier to change filename convention
+* fix : fix typo in grafanarmadillo.flow.Store : write_dasbaord -> write_dashboard
+
 v0.7.1 (2024-06-03)
 ------------------------------------------------------------